Playing with protobuf flags

+++ robustus/detail/install_protobuf.py
@@ -16,10 +16,15 @@ def install(robustus, requirement_specifier, rob_file, ignore_index):
     os.chdir(robustus.cache)
 
     install_dir = os.path.join(robustus.cache, 'protobuf-%s' % requirement_specifier.version)
-    if not os.path.isdir(install_dir) and not ignore_index:
+    
+    # always recompile protobuf because it has been cached buggy
+    # if not os.path.isdir(install_dir) and not ignore_index:
+    if True:
         archive_name = 'protobuf-%s.tar.gz' % requirement_specifier.version
-        subprocess.call(['wget', '-c', 'https://protobuf.googlecode.com/svn/rc/%s' % (archive_name,)])
-        subprocess.call(['tar', 'zxvf', archive_name])
+        run_shell(['wget', '-c', 'https://protobuf.googlecode.com/svn/rc/%s' % (archive_name,)],
+                  verbose=robustus.settings['verbosity'] >= 1)
+        run_shell(['tar', 'zxvf', archive_name],
+                  verbose=robustus.settings['verbosity'] >= 1)
 
         # move sources to a folder in order to use a clean name for installation
         src_dir = 'protobuf-%s' % requirement_specifier.version
@@ -29,9 +34,23 @@ def install(robustus, requirement_specifier, rob_file, ignore_index):
         os.chdir(src_dir)
         os.mkdir(install_dir)
 
-        subprocess.call(['./configure', '--disable-shared', '--prefix', install_dir])
-        subprocess.call('make', shell=True)
-        subprocess.call('make install', shell=True)
+        retcode = run_shell(['./configure', '--disable-shared',
+                             'CFLAGS=-fPIC',
+                             'CXXFLAGS=-fPIC',
+                             '--prefix', install_dir],
+                            verbose=robustus.settings['verbosity'] >= 1)
+
+        if retcode:
+            raise RequirementException('Failed to configure protobuf compilation')
+        retcode = run_shell('make', shell=True,
+                            verbose=robustus.settings['verbosity'] >= 1)
+        if retcode:
+            raise RequirementException('Failed compile protobuf')
+
+        retcode = run_shell('make install', shell=True)
+        if retcode:
+            raise RequirementException('Failed install protobuf')
+
         os.chdir(robustus.cache)
         shutil.rmtree(src_dir)
